TO BE SUBMITTED
Didn't have to use phaseR and Jacobian and Eigen functions. I've removed everything from phaseR. In order to plot the nullclines, I've defined the equations separately and used curve. Have to solve the Jacobian matrix and get Eigen values on paper but I've still included them in the code to give the answer some conclusion.